+++ rest_framework_json_api/exceptions.py
@@ -1,9 +1,16 @@
-import inspect
-from django.utils import six, encoding
+from django.conf import settings
 from django.utils.translation import ugettext_lazy as _
 from rest_framework import status, exceptions
 
-from rest_framework_json_api.utils import format_value
+from rest_framework_json_api import utils
+from rest_framework_json_api import renderers
+
+
+def rendered_with_json_api(view):
+    for renderer_class in getattr(view, 'renderer_classes', []):
+        if issubclass(renderer_class, renderers.JSONRenderer):
+            return True
+    return False
 
 
 def exception_handler(exc, context):
@@ -14,67 +21,28 @@ def exception_handler(exc, context):
     #
     # Also see: https://github.com/django-json-api/django-rest-framework-json-api/issues/158
     from rest_framework.views import exception_handler as drf_exception_handler
-    response = drf_exception_handler(exc, context)
 
+    # Render exception with DRF
+    response = drf_exception_handler(exc, context)
     if not response:
         return response
 
-    errors = []
-    # handle generic errors. ValidationError('test') in a view for example
-    if isinstance(response.data, list):
-        for message in response.data:
-            errors.append({
-                'detail': message,
-                'source': {
-                    'pointer': '/data',
-                },
-                'status': encoding.force_text(response.status_code),
-            })
-    # handle all errors thrown from serializers
-    else:
-        for field, error in response.data.items():
-            field = format_value(field)
-            pointer = '/data/attributes/{}'.format(field)
-            # see if they passed a dictionary to ValidationError manually
-            if isinstance(error, dict):
-                errors.append(error)
-            elif isinstance(error, six.string_types):
-                classes = inspect.getmembers(exceptions, inspect.isclass)
-                # DRF sets the `field` to 'detail' for its own exceptions
-                if isinstance(exc, tuple(x[1] for x in classes)):
-                    pointer = '/data'
-                errors.append({
-                    'detail': error,
-                    'source': {
-                        'pointer': pointer,
-                    },
-                    'status': encoding.force_text(response.status_code),
-                })
-            elif isinstance(error, list):
-                for message in error:
-                    errors.append({
-                        'detail': message,
-                        'source': {
-                            'pointer': pointer,
-                        },
-                        'status': encoding.force_text(response.status_code),
-                    })
-            else:
-                errors.append({
-                    'detail': error,
-                    'source': {
-                        'pointer': pointer,
-                    },
-                    'status': encoding.force_text(response.status_code),
-                })
+    # Use regular DRF format if not rendered by DRF JSON API and not uniform
+    is_json_api_view = rendered_with_json_api(context['view'])
+    is_uniform = getattr(settings, 'JSON_API_UNIFORM_EXCEPTIONS', False)
+    if not is_json_api_view and not is_uniform:
+        return response
 
+    # Convert to DRF JSON API error format
+    response = utils.format_drf_errors(response, context, exc)
 
-    context['view'].resource_name = 'errors'
-    response.data = errors
+    # Add top-level 'errors' object when not rendered by DRF JSON API
+    if not is_json_api_view:
+        response.data = utils.format_errors(response.data)
+
     return response
 
 
 class Conflict(exceptions.APIException):
     status_code = status.HTTP_409_CONFLICT
     default_detail = _('Conflict.')
-