+++ rest_framework_json_api/renderers.py
@@ -43,6 +43,9 @@ class JSONRenderer(renderers.JSONRenderer):
             # ID is always provided in the root of JSON API so remove it from attributes
             if field_name == 'id':
                 continue
+            # don't output a key for write only fields
+            if fields[field_name].write_only:
+                continue
             # Skip fields with relations
             if isinstance(field, (relations.RelatedField, relations.ManyRelatedField, BaseSerializer)):
                 continue
@@ -208,7 +211,7 @@ class JSONRenderer(renderers.JSONRenderer):
 
             if isinstance(field, ModelSerializer):
                 relation_model = field.Meta.model
-                relation_type = utils.format_relation_name(relation_model.__name__)
+                relation_type = utils.format_type(relation_model.__name__)
 
                 data.update({
                     field_name: {
@@ -248,7 +251,9 @@ class JSONRenderer(renderers.JSONRenderer):
                 included_resources.remove(field_name)
             except ValueError:
                 # Skip fields not in requested included resources
-                continue
+                # If no child field, directly continue with the next field
+                if field_name not in [node.split('.')[0] for node in included_resources]:
+                    continue
 
             try:
                 relation_instance_or_manager = getattr(resource_instance, field_name)
@@ -290,9 +295,13 @@ class JSONRenderer(renderers.JSONRenderer):
                     for position in range(len(serializer_data)):
                         serializer_resource = serializer_data[position]
                         nested_resource_instance = relation_queryset[position]
+                        resource_type = (
+                            relation_type or
+                            utils.get_resource_type_from_instance(nested_resource_instance)
+                        )
                         included_data.append(
                             JSONRenderer.build_json_resource_obj(
-                                serializer_fields, serializer_resource, nested_resource_instance, relation_type
+                                serializer_fields, serializer_resource, nested_resource_instance, resource_type
                             )
                         )
                         included_data.extend(
@@ -336,13 +345,18 @@ class JSONRenderer(renderers.JSONRenderer):
         return data
 
     @staticmethod
-    def extract_root_meta(serializer, resource, meta):
+    def extract_root_meta(serializer, resource):
+        many = False
+        if hasattr(serializer, 'child'):
+            many = True
+            serializer = serializer.child
+
+        data = {}
         if getattr(serializer, 'get_root_meta', None):
-            root_meta = serializer.get_root_meta(resource)
-            if root_meta:
-                assert isinstance(root_meta, dict), 'get_root_meta must return a dict'
-                meta.update(root_meta)
-        return meta
+            json_api_meta = serializer.get_root_meta(resource, many)
+            assert isinstance(json_api_meta, dict), 'get_root_meta must return a dict'
+            data.update(json_api_meta)
+        return data
 
     @staticmethod
     def build_json_resource_obj(fields, resource, resource_instance, resource_name):
@@ -374,11 +388,8 @@ class JSONRenderer(renderers.JSONRenderer):
         )
 
     def render_errors(self, data, accepted_media_type=None, renderer_context=None):
-        # Get the resource name.
-        if len(data) > 1 and isinstance(data, list):
-            data.sort(key=lambda x: x.get('source', {}).get('pointer', ''))
         return super(JSONRenderer, self).render(
-            {'errors': data}, accepted_media_type, renderer_context
+            utils.format_errors(data), accepted_media_type, renderer_context
         )
 
     def render(self, data, accepted_media_type=None, renderer_context=None):
@@ -410,6 +421,7 @@ class JSONRenderer(renderers.JSONRenderer):
         else:
             included_resources = list()
 
+        json_api_data = data
         json_api_included = list()
         # initialize json_api_meta with pagination meta or an empty dict
         json_api_meta = data.get('meta', {}) if isinstance(data, dict) else {}
@@ -419,51 +431,44 @@ class JSONRenderer(renderers.JSONRenderer):
         else:
             serializer_data = data
 
-        if hasattr(serializer_data, 'serializer') and getattr(serializer_data.serializer, 'many', False):
-            # The below is not true for non-paginated responses
-            # and isinstance(data, dict):
+        serializer = getattr(serializer_data, 'serializer', None)
 
-            # If detail view then json api spec expects dict, otherwise a list
-            # - http://jsonapi.org/format/#document-top-level
-            # The `results` key may be missing if unpaginated or an OPTIONS request
-
-            resource_serializer = serializer_data.serializer
+        if serializer is not None:
 
             # Get the serializer fields
-            fields = utils.get_serializer_fields(resource_serializer)
+            fields = utils.get_serializer_fields(serializer)
 
-            json_api_data = list()
-            for position in range(len(serializer_data)):
-                resource = serializer_data[position]  # Get current resource
-                resource_instance = resource_serializer.instance[position]  # Get current instance
+            # Extract root meta for any type of serializer
+            json_api_meta.update(self.extract_root_meta(serializer, serializer_data))
 
-                json_resource_obj = self.build_json_resource_obj(fields, resource, resource_instance, resource_name)
-                meta = self.extract_meta(resource_serializer, resource)
-                if meta:
-                    json_resource_obj.update({'meta': utils.format_keys(meta)})
-                json_api_meta = self.extract_root_meta(resource_serializer, resource, json_api_meta)
-                json_api_data.append(json_resource_obj)
+            if getattr(serializer, 'many', False):
+                json_api_data = list()
 
-                included = self.extract_included(fields, resource, resource_instance, included_resources)
-                if included:
-                    json_api_included.extend(included)
-        else:
-            # Check if data contains a serializer
-            if hasattr(data, 'serializer'):
-                fields = utils.get_serializer_fields(data.serializer)
-                resource_instance = data.serializer.instance
-                json_api_data = self.build_json_resource_obj(fields, data, resource_instance, resource_name)
+                for position in range(len(serializer_data)):
+                    resource = serializer_data[position]  # Get current resource
+                    resource_instance = serializer.instance[position]  # Get current instance
 
-                meta = self.extract_meta(data.serializer, data)
+                    json_resource_obj = self.build_json_resource_obj(fields, resource, resource_instance, resource_name)
+                    meta = self.extract_meta(serializer, resource)
+                    if meta:
+                        json_resource_obj.update({'meta': utils.format_keys(meta)})
+                    json_api_data.append(json_resource_obj)
+
+                    included = self.extract_included(fields, resource, resource_instance, included_resources)
+                    if included:
+                        json_api_included.extend(included)
+            else:
+                resource_instance = serializer.instance
+                json_api_data = self.build_json_resource_obj(fields, serializer_data, resource_instance, resource_name)
+
+                meta = self.extract_meta(serializer, serializer_data)
                 if meta:
                     json_api_data.update({'meta': utils.format_keys(meta)})
-                json_api_meta = self.extract_root_meta(data.serializer, data, json_api_meta)
 
-                included = self.extract_included(fields, data, resource_instance, included_resources)
+                included = self.extract_included(fields, serializer_data, resource_instance, included_resources)
                 if included:
                     json_api_included.extend(included)
-            else:
-                json_api_data = data
+
 
         # Make sure we render data in a specific order
         render_data = OrderedDict()